refactor: move analytics home sections into sites (#1598)
