+++ app/views/HomeView/ga2/homeView.tsx
@@ -1,6 +1,6 @@
-import { SectionAnalyticsAndData } from "@/components/Home/components/Section/components/ga2/SectionAnalyticsAndData/sectionAnalyticsAndData";
 import { SectionAssemblies } from "@/components/Home/components/Section/components/ga2/SectionAssemblies/sectionAssemblies";
 import { SectionHero } from "@/components/Home/components/Section/components/ga2/SectionHero/sectionHero";
+import { SectionAnalyticsAndData } from "@ga2/views/HomeView/components/SectionAnalyticsAndData/sectionAnalyticsAndData";
 import { SectionSubHero } from "@ga2/views/HomeView/components/SectionSubHero/sectionSubHero";
 import { Fragment, type JSX } from "react";
