+++ app/views/HomeView/homeView.tsx
@@ -1,5 +1,5 @@
-import { SectionAnalytics } from "@/components/Home/components/Section/components/SectionAnalytics/sectionAnalytics";
 import { SectionAssemblies } from "@/components/Home/components/Section/components/SectionAssemblies/sectionAssemblies";
+import { SectionAnalytics } from "@brc/views/HomeView/components/SectionAnalytics/sectionAnalytics";
 import { SectionHero } from "@brc/views/HomeView/components/SectionHero/sectionHero";
 import { SectionSubHero } from "@brc/views/HomeView/components/SectionSubHero/sectionSubHero";
 import { SectionWhitePapers } from "@brc/views/HomeView/components/SectionWhitePapers/sectionWhitePapers";
