+++ sites/brc-analytics/views/HomeView/components/SectionAnalytics/constants.ts
