+++ sites/brc-analytics/views/HomeView/components/SectionAnalytics/sectionAnalytics.styles.ts
