+++ sites/brc-analytics/views/HomeView/components/SectionAnalytics/sectionAnalytics.tsx
@@ -4,7 +4,7 @@ import {
   SectionTitle,
 } from "@repo/shared/views/HomeView/components/Section/section.styles";
 import { type JSX } from "react";
-import { ANALYTICS_TOOLS } from "./components/AnalyticsTools/common/constants";
+import { ANALYTICS_TOOLS } from "./constants";
 import { Headline, Section, SectionLayout } from "./sectionAnalytics.styles";
 
 export const SectionAnalytics = (): JSX.Element => {
